requirements.txt
# File Organizer - Complete Dependencies
# Install with: pip install -r requirements.txt --break-system-packages

# ===== CORE DEPENDENCIES (Already Working) =====
reportlab>=4.0.0
matplotlib>=3.7.0

# ===== SECURITY MODULES =====
cryptography>=41.0.0

# ===== MACHINE LEARNING & AI =====
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# ===== WEB & API (Optional - for Web Dashboard) =====
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# ===== ADVANCED ANALYTICS (Optional - for Interactive Dashboards) =====
plotly>=5.17.0
dash>=2.14.0

# ===== AUTOMATION (Optional - for File Watchers) =====
watchdog>=3.0.0
schedule>=1.2.0

# ===== CLOUD INTEGRATION (Optional - for Cloud Backup) =====
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
boto3>=1.28.0
dropbox>=11.36.0

# ===== MEDIA PROCESSING (Optional - for Photo/Video Organization) =====
Pillow>=10.0.0
# opencv-python>=4.8.0  # Uncomment if needed (large download)
# face-recognition>=1.3.0  # Uncomment if needed (requires dlib)

# ===== ENTERPRISE (Optional - for Corporate Features) =====
# ldap3>=2.9.0  # Uncomment for Active Directory
# celery>=5.3.0  # Uncomment for distributed processing
# redis>=5.0.0  # Uncomment for task queue

# ===== DEVELOPER TOOLS (Optional - for CLI) =====
click>=8.1.0

# ===== NOTES =====
# Use --break-system-packages flag on Linux/Mac if needed
# Comment out (add #) to any optional sections you don't want